UserSkill_ApI/.idea/requirements.txt
flask
flask_sqlalchemy
flask_restful
flask_httpauth
psycopg2-binary
pytest
python-decouple
